bae_academic_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai==1.3.7
pydantic==2.5.0
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
sqlalchemy==2.0.23
langgraph==0.0.26
aiosqlite==0.19.0
pandas==2.1.4 